+++ examples/factories/project-factory/main.tf
@@ -15,27 +15,26 @@
  */
 
 locals {
-  _gke_iam_hsau = try(var.vpc.gke_setup.enable_security_admin, false) ? {
-    "roles/container.hostServiceAgentUser" = [
-      "serviceAccount:${local.service_accounts_robots["container-engine"]}"
-  ] } : {}
-
-  _gke_iam_securityadmin = try(var.vpc.gke_setup.enable_security_admin, false) ? {
-    "roles/compute.securityAdmin" = [
-      "serviceAccount:${local.service_accounts_robots["container-engine"]}"
-  ] } : {}
+  # internal structures for group IAM bindings
   _group_iam = {
-    for r in local._group_iam_roles : r => [
-      for k, v in var.group_iam : "group:${k}" if try(index(v, r), null) != null
+    for r in local._group_iam_bindings : r => [
+      for k, v in var.group_iam :
+      "group:${k}" if try(index(v, r), null) != null
     ]
   }
-  _group_iam_roles = distinct(flatten(values(var.group_iam)))
+  _group_iam_bindings = distinct(flatten(values(var.group_iam)))
+  # internal structures for project service accounts IAM bindings
   _service_accounts_iam = {
-    for r in local._service_accounts_iam_roles : r => [
-      for k, v in var.service_accounts : "serviceAccount:${k}@${var.project_id}.iam.gserviceaccount.com" if try(index(v, r), null) != null
+    for r in local._service_accounts_iam_bindings : r => [
+      for k, v in var.service_accounts :
+      "serviceAccount:${k}@${var.project_id}.iam.gserviceaccount.com"
+      if try(index(v, r), null) != null
     ]
   }
-  _service_accounts_iam_roles = distinct(flatten(values(var.service_accounts)))
+  _service_accounts_iam_bindings = distinct(flatten(
+    values(var.service_accounts)
+  ))
+  # internal structures for project services
   _services = concat([
     "billingbudgets.googleapis.com",
     "essentialcontacts.googleapis.com"
@@ -44,64 +43,81 @@ locals {
     try(var.vpc.gke_setup, null) != null ? ["container.googleapis.com"] : [],
     var.vpc != null ? ["compute.googleapis.com"] : [],
   )
-  _services_iam_roles = distinct(flatten(values(var.services_iam)))
-  _services_iam = {
-    for r in local._services_iam_roles : r => [
-      for k, v in var.services_iam : "serviceAccount:${local.service_accounts_robots[k]}" if try(index(v, r), null) != null
+  # internal structures for service identity IAM bindings
+  _service_identities_roles = distinct(flatten(values(var.service_identities_iam)))
+  _service_identities_iam = {
+    for role in local._service_identities_roles : role => [
+      for service, roles in var.service_identities_iam :
+      "serviceAccount:${module.project.service_accounts.robots[service]}"
+      if contains(roles, role)
     ]
   }
-  billing_account_id = coalesce(var.billing_account_id, var.defaults.billing_account_id)
-  billing_alert      = var.billing_alert == null ? var.defaults.billing_alert : var.billing_alert
-  essential_contacts = concat(try(var.defaults.essential_contacts, []), var.essential_contacts)
+  # internal structure for Shared VPC service project IAM bindings
+  _vpc_subnet_bindings = (
+    local.vpc.subnets_iam == null || local.vpc.host_project == null
+    ? []
+    : flatten([
+      for subnet, members in local.vpc.subnets_iam : [
+        for member in members : {
+          region = split("/", subnet)[0]
+          subnet = split("/", subnet)[1]
+          member = member
+        }
+      ]
+    ])
+  )
+  # structures for billing id
+  billing_account_id = coalesce(
+    var.billing_account_id, try(var.defaults.billing_account_id, "")
+  )
+  billing_alert = (
+    var.billing_alert == null
+    ? try(var.defaults.billing_alert, null)
+    : var.billing_alert
+  )
+  # structure for essential contacts
+  essential_contacts = concat(
+    try(var.defaults.essential_contacts, []), var.essential_contacts
+  )
+  # structure that combines all authoritative IAM bindings
   iam = {
     for role in distinct(concat(
       keys(var.iam),
       keys(local._group_iam),
-      keys(local._gke_iam_hsau),
-      keys(local._gke_iam_securityadmin),
       keys(local._service_accounts_iam),
-      keys(local._services_iam),
+      keys(local._service_identities_iam),
     )) :
     role => concat(
       try(var.iam[role], []),
       try(local._group_iam[role], []),
-      try(local._gke_iam_hsau[role], []),
-      try(local._gke_iam_securityadmin[role], []),
       try(local._service_accounts_iam[role], []),
-      try(local._services_iam[role], []),
+      try(local._service_identities_iam[role], []),
     )
   }
-  labels = merge(coalesce(var.labels, {}), coalesce(var.defaults.labels, {}))
-  network_user_service_accounts = concat(
-    contains(local.services, "compute.googleapis.com") ? [
-      "serviceAccount:${local.service_accounts_robots.compute}"
-    ] : [],
-    contains(local.services, "container.googleapis.com") ? [
-      "serviceAccount:${local.service_accounts_robots.container-engine}",
-      "serviceAccount:${local.service_accounts.cloud_services}"
-    ] : [],
-  [])
+  # merge labels with defaults
+  labels = merge(
+    coalesce(var.labels, {}), coalesce(try(var.defaults.labels, {}), {})
+  )
+  # deduplicate services
   services = distinct(concat(var.services, local._services))
-  service_accounts_robots = {
-    for service, name in local.service_accounts_robot_services :
-    service => "${service == "bq" ? "bq" : "service"}-${module.project.number}@${name}.iam.gserviceaccount.com"
+  # structures for Shared VPC resources in host project
+  vpc = coalesce(var.vpc, {
+    host_project = null, gke_setup = null, subnets_iam = null
+  })
+  vpc_cloudservices = (
+    local.vpc_gke_service_agent ||
+    contains(var.services, "compute.googleapis.com")
+  )
+  vpc_gke_security_admin = coalesce(
+    try(local.vpc.gke_setup.enable_security_admin, null), false
+  )
+  vpc_gke_service_agent = coalesce(
+    try(local.vpc.gke_setup.enable_host_service_agent, null), false
+  )
+  vpc_subnet_bindings = {
+    for binding in local._vpc_subnet_bindings :
+    "${binding.subnet}:${binding.member}" => binding
   }
-  service_accounts_robot_services = {
-    cloudasset        = "gcp-sa-cloudasset"
-    cloudbuild        = "gcp-sa-cloudbuild"
-    compute           = "compute-system"
-    container-engine  = "container-engine-robot"
-    containerregistry = "containerregistry"
-    dataflow          = "dataflow-service-producer-prod"
-    dataproc          = "dataproc-accounts"
-    gae-flex          = "gae-api-prod"
-    gcf               = "gcf-admin-robot"
-    pubsub            = "gcp-sa-pubsub"
-    secretmanager     = "gcp-sa-secretmanager"
-    storage           = "gs-project-accounts"
-  }
-  vpc_host_project = try(var.vpc.host_project, var.defaults.vpc_host_project)
-  vpc_setup        = var.vpc != null
 }
 
 module "billing-alert" {
@@ -134,6 +150,7 @@ module "project" {
   source                     = "../../../modules/project"
   billing_account            = local.billing_account_id
   name                       = var.project_id
+  prefix                     = var.prefix
   contacts                   = { for c in local.essential_contacts : c => ["ALL"] }
   iam                        = local.iam
   labels                     = local.labels
@@ -142,9 +159,21 @@ module "project" {
   policy_list                = try(var.org_policies.policy_list, {})
   service_encryption_key_ids = var.kms_service_agents
   services                   = local.services
-  shared_vpc_service_config = {
-    attach       = local.vpc_setup
-    host_project = local.vpc_host_project
+  shared_vpc_service_config = var.vpc == null ? null : {
+    host_project = local.vpc.host_project
+    # these are non-authoritative
+    service_identity_iam = {
+      "roles/compute.networkUser" = compact([
+        local.vpc_gke_service_agent ? "container-engine" : null,
+        local.vpc_cloudservices ? "cloudservices" : null
+      ])
+      "roles/compute.securityAdmin" = compact([
+        local.vpc_gke_security_admin ? "container-engine" : null,
+      ])
+      "roles/container.hostServiceAgentUser" = compact([
+        local.vpc_gke_service_agent ? "container-engine" : null
+      ])
+    }
   }
 }
 
@@ -155,11 +184,11 @@ module "service-accounts" {
   project_id = module.project.project_id
 }
 
-resource "google_compute_subnetwork_iam_binding" "binding" {
-  for_each   = local.vpc_setup ? coalesce(var.vpc.subnets_iam, {}) : {}
-  project    = local.vpc_host_project
-  subnetwork = "projects/${local.vpc_host_project}/regions/${split("/", each.key)[0]}/subnetworks/${split("/", each.key)[1]}"
-  region     = split("/", each.key)[0]
+resource "google_compute_subnetwork_iam_member" "default" {
+  for_each   = local.vpc_subnet_bindings
+  project    = local.vpc.host_project
+  subnetwork = "projects/${local.vpc.host_project}/regions/${each.value.region}/subnetworks/${each.value.subnet}"
+  region     = each.value.region
   role       = "roles/compute.networkUser"
-  members    = concat(each.value, local.network_user_service_accounts)
+  member     = each.value.member
 }