+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
 -  **incompatible change** removed `iam` key from logging sink configuration in the `project` and `organization` modules
+- remove GCS to BQ with Dataflow example, replace by GCS to BQ with least privileges
+- the `net-vpc` and `project` modules now use the beta provider for shared VPC-related resources
+- new `iot-core` module
+- **incompatible change** the variables for host and service Shared VPCs have changed in the project module
+- **incompatible change** the variable for service identities IAM has changed in the project factory
 
 ## [13.0.0] - 2022-01-27
 
@@ -15,7 +21,6 @@ All notable changes to this project will be documented in this file.
 - support service dependencies for crypto key bindings in project module
 - refactor project module in multiple files
 - add support for per-file option overrides to tfdoc
-- the `net-vpc` and `project` modules now use the beta provider for shared VPC-related resources
 
 ## [12.0.0] - 2022-01-11